code refactor to share tree functions between results provider
